[2491] build: remove docker-compose.init.sh, sqlite3.db
Instead of requiring `docker-compose.init.sh` to be run to create the
bind mount for `sqlite3.db`, leave `sqlite3.db` to be auto-created
within the container. This simplifies getting up and running.

If you want the database to persist beyond the life of the container,
either use a bind mount or a volume. Environment variable `SQLITE_FILE`
is available for this purpose, and it defaults to `sqlite3.db`.